Get a decent gif generation utility running

+++ screencast.sh
@@ -1,16 +1,25 @@
-#!/bin/bash -x
+#!/bin/bash
 
 SCREENCAST_NAME_PATH=~/.screencasting.file.path
 OUTPUT_FILE=$(<$SCREENCAST_NAME_PATH)
+
 if [ !$OUTPUT_FILE ]
 then
-	OUTPUT_FILE="$(date +'%Y-%m-%dT%I-%M%p').mov"
+	OUTPUT_FILE="$(date +'%Y-%m-%dT%I-%M%p')"
 fi
-DISPLAY=:0.0 /usr/bin/i3-nagbar -m "Recording to $OUTPUT_FILE..." -t warning &
+MOV_FILE=$OUTPUT_FILE.mov
+GIF_FILE=$OUTPUT_FILE.gif
+FINAL_OUTPUT=$MOV_FILE
+
+if [[ "$@" =~ "--to-gif" ]]
+then
+	FINAL_OUTPUT=$GIF_FILE
+fi
+DISPLAY=:0.0 /usr/bin/i3-nagbar -m "Recording to $FINAL_OUTPUT..." -t warning &
 
 INRES="1920x1200" # input resolution
 OUTRES="1920x1200"
-FPS="25" # target FPS
+FPS="15" # target FPS
 THREADS="2"
 QUALITY="medium" #x264 preset: ultrafast,superfast,veryfast,faster,fast,medium
 AUDIO_RATE="44100"
@@ -20,14 +29,23 @@ ffmpeg  -nostdin \
 	-thread_queue_size 512 -f alsa -i pulse -ac 2 -ar $AUDIO_RATE \
 	-vcodec libx264 -preset $QUALITY -s "$OUTRES" -pix_fmt yuv420p -tune film \
 	-acodec libmp3lame -ar 44100 -ab 64k -threads $THREADS -strict normal \
-	-f mov file:$OUTPUT_FILE
+	-f mov file:$MOV_FILE
 
-if [ $? ]
+FFMPEG_STATUS_CODE=$?
+pkill i3-nagbar
+# 255 is used when ffmpeg is signalled via SIGTERM and then exits normally
+if [ $FFMPEG_STATUS_CODE ] && [ $FFMPEG_STATUS_CODE != '255' ]
 then
-	pkill i3-nagbar
-	if [ "$1" == "--from-toggle" ]
+	if [[ "$@" =~ "--from-toggle" ]]
 	then
 		rm $SCREENCAST_NAME_PATH
 		DISPLAY=:0.0 /usr/bin/i3-nagbar -m "Failed to record, debug with \`screencast.sh\`" -t warning &
 	fi
+else
+	if [[ "$@" =~ "--to-gif" ]]
+	then
+		ffmpeg -i $MOV_FILE $GIF_FILE
+		gifsicle -O2 $GIF_FILE -o $GIF_FILE
+	fi
+
 fi
